requirements.txt
torch
torchvision
argparse
wget
kaggle
pillow
